also update android to latest rc rn
