+++ android/app/src/main/java/com/pingpingnative/MainActivity.java
@@ -18,16 +18,6 @@ public class MainActivity extends ReactActivity {
         };
     }
 
-	/**
-	* Returns the instance of the {@link ReactActivityDelegate}. There the RootView is created and
-	* you can specify the renderer you wish to use - the new renderer (Fabric) or the old renderer
-	* (Paper).
-	*/
-	@Override
-	protected ReactActivityDelegate createReactActivityDelegate() {
-	return new MainActivityDelegate(this, getMainComponentName());
-	}
-
     /**
      * Returns the name of the main component registered from JavaScript. This is used to schedule
      * rendering of the component.
