+++ android/app/src/main/java/com/pingpingnative/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate.java
@@ -1,35 +1,4 @@
-package com.pingpingNative.newarchitecture.components;
-import com.facebook.jni.HybridData;
-import com.facebook.proguard.annotations.DoNotStrip;
-import com.facebook.react.fabric.ComponentFactory;
-import com.facebook.soloader.SoLoader;
-/**
- * Class responsible to load the custom Fabric Components. This class has native methods and needs a
- * corresponding C++ implementation/header file to work correctly (already placed inside the jni/
- * folder for you).
- *
- * <p>Please note that this class is used ONLY if you opt-in for the New Architecture (see the
- * `newArchEnabled` property). Is ignored otherwise.
- */
-@DoNotStrip
-public class MainComponentsRegistry {
-  static {
-    SoLoader.loadLibrary("fabricjni");
-  }
-  @DoNotStrip private final HybridData mHybridData;
-  @DoNotStrip
-  private native HybridData initHybrid(ComponentFactory componentFactory);
-  @DoNotStrip
-  private MainComponentsRegistry(ComponentFactory componentFactory) {
-    mHybridData = initHybrid(componentFactory);
-  }
-  @DoNotStrip
-  public static MainComponentsRegistry register(ComponentFactory componentFactory) {
-    return new MainComponentsRegistry(componentFactory);
-  }
-}
-android/app/src/main/java/com/pingpingNative/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate.java ADDED1 hidden comment
-package com.pingpingNative.newarchitecture.modules;
+package com.pingpingnative.newarchitecture.modules;
 import com.facebook.jni.HybridData;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.ReactPackageTurboModuleManagerDelegate;
@@ -64,7 +33,7 @@ public class MainApplicationTurboModuleManagerDelegate
     if (!sIsSoLibraryLoaded) {
       // If you change the name of your application .so file in the Android.mk file,
       // make sure you update the name here as well.
-      SoLoader.loadLibrary("pingpingNative_appmodules");
+      SoLoader.loadLibrary("pingpingnative_appmodules");
       sIsSoLibraryLoaded = true;
     }
   }
